+++ src/workflows/trades.ts
@@ -1,7 +1,7 @@
-import { Signer } from '@ethersproject/abstract-signer';
-import { StarkWallet } from '../types';
+import { L1Signer, L2Signer, StarkWallet } from '../types';
 import { GetSignableTradeRequest, TradesApi } from '../api';
 import { serializeSignature, sign, signRaw } from '../utils';
+import { Signer } from 'ethers';
 
 export async function createTradeWorkflow(
   signer: Signer,
@@ -10,7 +10,7 @@ export async function createTradeWorkflow(
   tradesApi: TradesApi,
 ) {
   // Obtain Ethereum Address from signer
-  const ethAddress = (await signer.getAddress());
+  const ethAddress = await signer.getAddress();
 
   const signableResult = await tradesApi.getSignableTrade({
     getSignableTradeRequest: {
@@ -30,6 +30,53 @@ export async function createTradeWorkflow(
   const starkSignature = serializeSignature(
     sign(starkWallet.starkKeyPair, payloadHash),
   );
+
+  const createTradeResponse = await tradesApi.createTrade({
+    createTradeRequest: {
+      amount_buy: signableResult.data.amount_buy,
+      amount_sell: signableResult.data.amount_sell,
+      asset_id_buy: signableResult.data.asset_id_buy,
+      asset_id_sell: signableResult.data.asset_id_sell,
+      expiration_timestamp: signableResult.data.expiration_timestamp,
+      fee_info: signableResult.data.fee_info,
+      fees: request.fees,
+      include_fees: true,
+      nonce: signableResult.data.nonce,
+      order_id: request.order_id,
+      stark_key: signableResult.data.stark_key,
+      vault_id_buy: signableResult.data.vault_id_buy,
+      vault_id_sell: signableResult.data.vault_id_sell,
+      stark_signature: starkSignature,
+    },
+    xImxEthAddress: ethAddress,
+    xImxEthSignature: ethSignature,
+  });
+
+  return createTradeResponse.data;
+}
+
+export async function createTradeWorkflowWithSigner(
+  l1Signer: L1Signer,
+  l2Signer: L2Signer,
+  request: GetSignableTradeRequest,
+  tradesApi: TradesApi,
+) {
+  const ethAddress = await l1Signer.getAddress();
+
+  const signableResult = await tradesApi.getSignableTrade({
+    getSignableTradeRequest: {
+      user: ethAddress,
+      order_id: request.order_id,
+      fees: request.fees,
+    },
+  });
+
+  const { signable_message: signableMessage, payload_hash: payloadHash } =
+    signableResult.data;
+
+  const ethSignature = await signRaw(signableMessage, l1Signer);
+
+  const starkSignature = await l2Signer.signMessage(payloadHash);
 
   const createTradeResponse = await tradesApi.createTrade({
     createTradeRequest: {