+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added `BaseSigner`, a default implementation of the Stark L2Signer interface
+- Added `createTradeWithSigner` function to enable creating trade workflow with l2signer
+
+### Changed
+- mark `createTrade` as deprecated 
+
 ## [0.3.1] - 2022-07-01
 
 ### Fixed